Crate nam_redjubjub

A minimal RedJubjub implementation for use in Zebra.

Two parameterizations of RedJubjub are used in Zcash, one for BindingSig and one for SpendAuthSig. This library distinguishes these in the type system, using the sealed SigType trait as a type-level enum.

In addition to the Signature, SigningKey, VerificationKey types, the library also provides VerificationKeyBytes, a refinement of a [u8; 32] indicating that bytes represent an encoding of a RedJubjub verification key. This allows the VerificationKey type to cache verification checks related to the verification key encoding.

Examples

Creating a BindingSig, serializing and deserializing it, and verifying the signature:

use rand::thread_rng;
use redjubjub::*;

let msg = b"Hello!";

// Generate a secret key and sign the message
let sk = SigningKey::<Binding>::new(thread_rng());
let sig = sk.sign(thread_rng(), msg);

// Types can be converted to raw byte arrays using From/Into
let sig_bytes: [u8; 64] = sig.into();
let pk_bytes: [u8; 32] = VerificationKey::from(&sk).into();

// Deserialize and verify the signature.
let sig: Signature<Binding> = sig_bytes.into();
assert!(
    VerificationKey::try_from(pk_bytes)
        .and_then(|pk| pk.verify(msg, &sig))
        .is_ok()
);

docs

cargo doc --features "nightly" --open

Modules

batch

Performs batch RedJubjub signature verification.

Structs

Signature

A RedJubJub signature.

SigningKey

A RedJubJub signing key.

VerificationKey

A valid RedJubJub verification key.

VerificationKeyBytes

A refinement type for [u8; 32] indicating that the bytes represent an encoding of a RedJubJub verification key.

Enums

Binding

A type variable corresponding to Zcash’s BindingSig.

Error

An error related to RedJubJub signatures.

SpendAuth

A type variable corresponding to Zcash’s SpendAuthSig.

Traits

SigType

Abstracts over different RedJubJub parameter choices, Binding and SpendAuth.

Type Aliases

Randomizer

An element of the JubJub scalar field used for randomization of public and secret keys.